ProyectoRFID/requirements.txt
azure-functions==1.16.0
azure-cosmos==4.6.0